Add support for Varnish 6.2.0
The code generated by vmodtool.py is now slightly different to avoid
symbol collisions under certain circumstances. To support both 6.2.0
and 6.0.0 we can emulate the compatibility macros generated by newer
vmodtool.

One exception is the handling of the event function. It is renamed now
because the notion of an "init" function is a remnant of the Varnish 3
days and we can make a dedicated macro on the spot similar to what
vmodtool did for the rest.

Fixes #29
Closes #30

Signed-off-by: Dridi Boukelmoune <dridi@varnish-software.com>